requirements.txt
Pillow
pillow-heif
tkinterdnd2
